requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==3.2.12
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4